Ignore add-ons with weird compatibility info when generating compatibility report
